separate player and enemy entities into own modules
