+++ src/main.rs
@@ -22,7 +22,7 @@ use crate::{
         state::GameState,
     },
     core::CorePlugin,
-    ships::PlayerPlugin,
+    ships::ShipsPlugin,
     ui::UiPlugin,
     world::WorldPlugin,
 };
@@ -52,7 +52,7 @@ fn main() {
             .set(ImagePlugin::default_nearest()),
     );
 
-    app.add_plugins((CorePlugin, PlayerPlugin, WorldPlugin, UiPlugin));
+    app.add_plugins((CorePlugin, ShipsPlugin, WorldPlugin, UiPlugin));
 
     #[cfg(debug_assertions)]
     app.add_plugins((
