+++ src/ships/mod.rs
@@ -2,16 +2,19 @@ use bevy::prelude::*;
 use bevy_rapier2d::prelude::*;
 
 pub mod dynamic_orbit;
+pub mod enemy;
+pub mod player;
 pub mod ship;
 
-use self::{dynamic_orbit::DynamicOrbitPlugin, ship::ShipPlugin};
+use self::{dynamic_orbit::DynamicOrbitPlugin, enemy::EnemyPlugin, player::PlayerPlugin};
 
-pub struct PlayerPlugin;
-impl Plugin for PlayerPlugin {
+pub struct ShipsPlugin;
+impl Plugin for ShipsPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins((
             RapierPhysicsPlugin::<NoUserData>::pixels_per_meter(1.0),
-            ShipPlugin,
+            PlayerPlugin,
+            EnemyPlugin,
             DynamicOrbitPlugin,
         ));
